// types/profilesdefs.ts
export interface Lootlist {
  itemType: string;
  itemGuid: string;
  itemProfile: string;
  quantity: number;
}